backend/requirements.txt
# FastAPI & Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
asyncpg>=0.29.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# pgvector support
pgvector>=0.2.4

# OCR 관련
easyocr>=1.7.1
opencv-python>=4.8.0
Pillow>=10.0.0

# QR 코드 디코딩
pyzbar>=0.1.9

# URL 처리 및 HTTP 요청
requests>=2.31.0

# 머신러닝 및 데이터 처리
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# 파일 업로드 처리
python-multipart>=0.0.6

# 유틸리티
tqdm>=4.66.0

# WHOIS 정보 조회
python-whois>=0.8.0
